Adding Sort panel and Finishing with Selection panel.
Selection panel has a series of buttons that each on follow different method for choosing a domain for number selection.
Sort panel is the main one that use a bunch of buttons to invoke different methods of sorting, which at this moment is not complete and only displays the random numbers that are generated. all of it happened in Sample scene and sort manager class.