Adding Number Prefab, EventManager, Numb class, Numb Selection class.
So there is this class named Numb which is attached to Number prefab.
Event Manager is a static class which is public and we use it to send numbers and object between managers like sort and selection manager.
Numb Selection is a struct of what we need from a selecting the right domain point of view.